Dev/Docker/docker-status: clickable entries
This commit gives each menu entry some clickable interactivity:

Docker machines which are on (and highlighted in green) - when
clicked - will be shut down (and subsequently all of their containers
will be stopped). If they are off (the ones highlighted in red) they
will be booted up.

The containers listed per machine now use tree like UTF-8 symbols to
better show their connection to their respective machine. In addition,
when clicked, they will be shut down. There is no concept of a
"non-running" container within this plugin, nor really with Docker, and
so there isn't an option to start containers up.

In addition to this, a small change has been made to the icon -
`dropdown=false` - to prevent the anchor icon showing up in the menus.