Webpack Build Status Plugin
This plugin, in combination with [BitBar Webpack Status Plugin(https://github.com/Fire-Dragon-DoL/bitbar-webpack-status-plugin) allows to display the current status of webpack build (success/failure/in progress). Very helpful if you are checking the browser right after saving a file and don't want to go back to the terminal to see when the compilation succeeds (Typescript is quite slow).