Fix a couple issues with npm/apm parsing (#480)
* Fix a couple issues with npm/apm parsing

* Wanted not current

* Proper support for system ruby
Requires sudo in order to do anything.  I recommend doing something like
`sudo dseditgroup -o edit -a  -t user wheel`
And then do `visudo` to make it so the `wheel` group does not require
a password.  (There is a line already there for it, you just need to
uncomment it and save.)

* Linked vs link