Added a new plugin vnc_short.
A simple plugin used a lot in our office to conveniently connect to
other computer. First block defines Menu bar icon. The next four
defines the VNC shortcuts. Change the VNC address to your targets IP
and port (5900 is default in OSX). Copy and paste to the desired
amount. Custom images can be encoded. Google base64 image encoding.
The recommended size for images in the status bar is 16x16.