Plugin Contribution - ISS Pass Finder (#547)
* Plugin Contribution - ISS Pass Finder

A script I adapted to display the upcoming passes of the International
Space Station over your location. Wasn’t quite sure where it belonged
but put it in lifestyle.

* Fixing issues from Travis

I was given a log of things to change and I believe I have fixed all
addressable issues. Travis was unhappy with my use of cat but cat is
necessary and the alternatives suggested broke my script

* Last correction I can make

This, as far as I can tell, is the last error that isn’t related to
“Useless cat”. I do not believe I can make the changes wanted around
those errors

* Trying to resolve useless cat issue

* Removed unnecessary lines. More Useless cat

* Removed single whitespace line