Add plugin to delete VSCode temp files
Visual Studio Code leaves a lot of temporary files around after running code and creates new ones anytime code is run, even if the old temp files are still there and the code is untouched. This deletes all files in the folder it is pointed to that start with 'temp'.