Add support for lua plugins + network bandwidth plugin in lua (#1084)
* Show total bandwidth for specified ifaces with lua

* Fix luacheck issues

* Add support for lua plugins

* Brew install luarocks instead of lua

We need luarocks to install luacheck for linting lua plugins. Brew
installing luarocks will automatically pull in a suitable version of
lua.

* Fix bitbar metadata

There was a typo in the github author, and the bitbar.desc was not
correctly recognized, probably because too long over multiple lines.

* Monitor all active interfaces

This does not require any user configuration.