Add ability to specify a title for copy text
I find it more useful to be able to name the copied text item rather than just have the full-text content appear in the menu. Maybe it's a password I want to hide or a long command I don't want to display. This merge request delimits the title and content.