Issue #1468 - Now Playing opens Podcasts in Catalina
Assume we are on Catalina or higher to start and not show an iTunes menu item. It appears "iTunes" is an alias under the hood for the new "Podcasts" app. Detect if we are on a pre-Catalina version of OS X and adjust the menu to have iTunes app instead of Music app.

Tried having "Podcasts" in the menu, but the app overall does not behave in the same way as the others. The check for the player state throws an error. Rather than dive down that rabbit hole, fix this bug and address Podcasts some other day.