Fix test script to only check actual file type, not charsets
The script would fail when a webserver decides that it has to append
a charset to an image content-type as it checks if the raw header is
in `allowed_image_content_types`; however that would fail if the header
does also include the charset. urllib provides a primitive that solves
this issue by always returning just the content type, not the encoding.