Update metadata
Sorry, forgot to update the version number of the plugin. Given that it is a major update to the config section implementation and some behavior fixes, I chose to up the major number.

I also added more dependencies but did not add the dependencies that I believe came preinstalled with bash like sed or printf.

I changed the image host for the screenshots to my own personal server (which I hope is more reliable and won't disappear on me).

I also renamed the repository hosting only this plugin to xbar and updated the url for the about page for this plugin.

Finally, I changed the variable descriptions to be more clear.