Final Commit
Tidies up the code a little - makes font size configurable and adds in current usage from the inverter and from the grid. This is about as far as this plugin can go for now without adding in EV Charging. 

NOTE: Will revisit when I have an electric car - already have the inverter.