Update install function & output generation
Starting with python version 3.12 homebrew installs of python are marked
as externally managed and require the '--break-system-packages' flag in
order to install extra modules. Try installing without the flag first in
the hopes that someday this will run in a virtual env, but if/when that
fails try again with the flag.

Update message output to put all messages in an array to ensure that the
title info remains on the macOs menu bar and error or other info is
displayed in the drop down menu.