Fix screenshot, use proper fetching (#2093)
Noticed thes screenshot here (and in several other places) is broken
because of imgur's http -> https redirect breaking whatever process is
loading them in. Instead, this moves it within the repository and uses
it from there.

Also avoids using the `open(url)` pattern because it tries to run it at
the Ruby kernel level rather than in a safer spot.
[ref](https://www.rubydoc.info/gems/rubocop/RuboCop/Cop/Security/Open)