Create NetStatus_ChineseUser.1s.sh (#2117)
Monitors internet connectivity and latency to key websites, displaying real-time status in the menu bar. Internet environment in China is complicated because of the "Great Fire Wall", so I created a script to monitor network status especially for this environment that can monitor the status of VPN and show different status of the internet with/without VPN, or in other words, if blocked sites are available and if the internet is simply down.